+++ Editor/Drawers/AttributeDrawers/GUIDAttributeDrawer.cs
@@ -4,8 +4,8 @@ using UnityEngine;
 
 namespace Treasured.UnitySdk.Editor
 {
-    [CustomPropertyDrawer(typeof(UniqueIdAttribute))]
-    public class UniqueIdAttributeDrawer : PropertyDrawer
+    [CustomPropertyDrawer(typeof(GUIDAttribute))]
+    public class GUIDAttributeDrawer : PropertyDrawer
     {
         public override void OnGUI(Rect position, SerializedProperty property, GUIContent label)
         {
@@ -14,7 +14,7 @@ namespace Treasured.UnitySdk.Editor
             {
                 using (new EditorGUI.DisabledGroupScope(true))
                 {
-                    EditorGUI.TextField(new Rect(position.x, position.y, position.width - 22, position.height), label, property.stringValue);
+                    EditorGUI.PropertyField(new Rect(position.x, position.y, position.width - 22, position.height), property, label);
                 }
                 if (GUI.Button(new Rect(position.xMax - 20, position.y, 20, position.height), EditorGUIUtility.TrIconContent("Refresh", "Regenerate Id")))
                 {