+++ Runtime/Component/IDataComponent.cs
@@ -1,9 +1,0 @@
-﻿namespace Treasured.UnitySdk
-{
-    public interface IDataComponent<T> where T : TreasuredObjectData
-    {
-        T Data { get; }
-
-        void BindData(T data);
-    }
-}