+++ Runtime/Component/Interactable.cs
@@ -5,7 +5,7 @@ using UnityEngine;
 namespace Treasured.UnitySdk
 {
     [AddComponentMenu("Treasured/Interactable")]
-    public sealed class Interactable : TreasuredObject, IDataComponent<InteractableData>
+    public sealed class Interactable : TreasuredObject
     {
         [Obsolete]
         [SerializeField]
@@ -19,31 +19,6 @@ namespace Treasured.UnitySdk
             {
                 return _data;
             }
-        }
-
-        [JsonIgnore]
-        InteractableData IDataComponent<InteractableData>.Data => _data;
-
-        public void BindData(InteractableData data)
-        {
-            gameObject.name = data.Name;
-            gameObject.transform.position = data.Transform.Position;
-            gameObject.transform.eulerAngles = data.Transform.Rotation;
-            BoxCollider.center = data.Hitbox.Center - data.Transform.Position;
-            BoxCollider.size = data.Hitbox.Size;
-            _data = data;
-        }
-
-        private Interactable() { }
-
-        protected override void OnEnable()
-        {
-            base.OnEnable();
-        }
-
-        void Reset()
-        {
-            _data?.Validate();
         }
     }
 }